requirements.txt
streamlit>=1.34.0
crewai==0.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
groq>=0.4.0
langchain==0.0.267
langchain-core==0.0.27
langchain-community==0.0.13
onnxruntime>=1.16.3
sentence-transformers>=2.2.2
pysqlite3-binary
types-protobuf
rich>=13.7.0,<14.0.0
chromadb==0.4.15
embedchain==0.1.60 